packages/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fractary-codex"
version = "0.1.0"
description = "Knowledge infrastructure SDK for AI agents - universal references, multi-tier caching, storage abstraction"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Fractary Engineering", email = "engineering@fractary.com" }
]
keywords = [
    "codex",
    "knowledge-management",
    "ai-agents",
    "mcp",
    "model-context-protocol",
    "caching",
    "storage",
    "async",
    "fractary"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "pyyaml>=6.0",
    "aiohttp>=3.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.1",
    "mypy>=1.8",
    "ruff>=0.1",
    "types-PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/fractary/codex"
Documentation = "https://github.com/fractary/codex#readme"
Repository = "https://github.com/fractary/codex.git"
Issues = "https://github.com/fractary/codex/issues"

[tool.hatch.build.targets.wheel]
packages = ["fractary_codex"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
target-version = "py39"
line-length = 88
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["fractary_codex"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
